+++ jest.config.mjs
@@ -1,9 +1,7 @@
-'use strict';
-
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
 
-module.exports = {
+export default {
   testEnvironment: 'jest-environment-jsdom-sixteen',
   preset: 'ts-jest',
   globals: {
@@ -14,12 +12,11 @@ module.exports = {
     }
   },
   coverageProvider: 'v8',
-  collectCoverage: process.env.CI === 'true',
   collectCoverageFrom: [
     'src/**/*.{ts,tsx}'
   ],
   coverageReporters: [
-    'cobertura'
+    'text'
   ],
   restoreMocks: true,
   moduleNameMapper: {