+++ src/Canvas.tsx
@@ -5,48 +5,48 @@ import EventBus from './EventBus';
 import InteractionMasks from './masks/InteractionMasks';
 import { DataGridProps } from './DataGrid';
 import RowRenderer from './RowRenderer';
-import SummaryRowRenderer from './SummaryRowRenderer';
+import SummaryRow from './SummaryRow';
 import { getColumnScrollPosition, getScrollbarSize, isPositionStickySupported, getVerticalRangeToRender, assertIsValidKey } from './utils';
 
-type SharedDataGridProps<R, K extends keyof R> = Pick<DataGridProps<R, K>,
-| 'rows'
-| 'rowRenderer'
-| 'rowGroupRenderer'
-| 'selectedRows'
-| 'summaryRows'
-| 'onCheckCellIsEditable'
-| 'onSelectedCellChange'
-| 'onSelectedCellRangeChange'
-| 'onRowClick'
-| 'onRowExpandToggle'
-| 'onSelectedRowsChange'
-| 'rowKey'
-> & Required<Pick<DataGridProps<R, K>,
-| 'enableCellAutoFocus'
-| 'enableCellCopyPaste'
-| 'enableCellDragAndDrop'
-| 'rowHeight'
-| 'cellNavigationMode'
-| 'editorPortalTarget'
-| 'onRowsUpdate'
+type SharedDataGridProps<R, K extends keyof R, SR> = Pick<DataGridProps<R, K, SR>,
+  | 'rows'
+  | 'rowRenderer'
+  | 'rowGroupRenderer'
+  | 'selectedRows'
+  | 'summaryRows'
+  | 'onCheckCellIsEditable'
+  | 'onSelectedCellChange'
+  | 'onSelectedCellRangeChange'
+  | 'onRowClick'
+  | 'onRowExpandToggle'
+  | 'onSelectedRowsChange'
+  | 'rowKey'
+> & Required<Pick<DataGridProps<R, K, SR>,
+  | 'enableCellAutoFocus'
+  | 'enableCellCopyPaste'
+  | 'enableCellDragAndDrop'
+  | 'rowHeight'
+  | 'cellNavigationMode'
+  | 'editorPortalTarget'
+  | 'onRowsUpdate'
 >>;
 
-export interface CanvasProps<R, K extends keyof R> extends SharedDataGridProps<R, K> {
-  columnMetrics: ColumnMetrics<R>;
-  viewportColumns: readonly CalculatedColumn<R>[];
+export interface CanvasProps<R, K extends keyof R, SR> extends SharedDataGridProps<R, K, SR> {
+  columnMetrics: ColumnMetrics<R, SR>;
+  viewportColumns: readonly CalculatedColumn<R, SR>[];
   height: number;
   scrollLeft: number;
-  onScroll(position: ScrollPosition): void;
+  onScroll: (position: ScrollPosition) => void;
 }
 
 export interface CanvasHandle {
-  scrollToColumn(colIdx: number): void;
-  scrollToRow(rowIdx: number): void;
-  selectCell(position: Position, openEditor?: boolean): void;
-  openCellEditor(rowIdx: number, colIdx: number): void;
+  scrollToColumn: (colIdx: number) => void;
+  scrollToRow: (rowIdx: number) => void;
+  selectCell: (position: Position, openEditor?: boolean) => void;
+  openCellEditor: (rowIdx: number, colIdx: number) => void;
 }
 
-function Canvas<R, K extends keyof R>({
+function Canvas<R, K extends keyof R, SR>({
   columnMetrics,
   viewportColumns,
   height,
@@ -59,7 +59,7 @@ function Canvas<R, K extends keyof R>({
   selectedRows,
   onSelectedRowsChange,
   ...props
-}: CanvasProps<R, K>, ref: React.Ref<CanvasHandle>) {
+}: CanvasProps<R, K, SR>, ref: React.Ref<CanvasHandle>) {
   const [eventBus] = useState(() => new EventBus());
   const [scrollTop, setScrollTop] = useState(0);
   const canvasRef = useRef<HTMLDivElement>(null);
@@ -193,7 +193,7 @@ function Canvas<R, K extends keyof R>({
       }
 
       rowElements.push(
-        <RowRenderer<R>
+        <RowRenderer<R, SR>
           key={key}
           rowIdx={rowIdx}
           row={row}
@@ -218,18 +218,15 @@ function Canvas<R, K extends keyof R>({
   const summary = summaryRows && summaryRows.length > 0 && (
     <div ref={summaryRef} className="rdg-summary">
       {summaryRows.map((row, rowIdx) => (
-        <SummaryRowRenderer<R>
+        <SummaryRow<R, SR>
           key={rowIdx}
           rowIdx={rowIdx}
           row={row}
           width={columnMetrics.totalColumnWidth + getScrollbarSize()}
           height={rowHeight}
           viewportColumns={viewportColumns}
-          isRowSelected={false}
           lastFrozenColumnIndex={columnMetrics.lastFrozenColumnIndex}
           scrollLeft={nonStickyScrollLeft}
-          isSummaryRow
-          eventBus={eventBus}
         />
       ))}
     </div>
@@ -243,7 +240,7 @@ function Canvas<R, K extends keyof R>({
         ref={canvasRef}
         onScroll={handleScroll}
       >
-        <InteractionMasks<R>
+        <InteractionMasks<R, SR>
           rows={rows}
           rowHeight={rowHeight}
           columns={columns}
@@ -281,4 +278,4 @@ function Canvas<R, K extends keyof R>({
 
 export default forwardRef(
   Canvas as React.RefForwardingComponent<CanvasHandle>
-) as <R, K extends keyof R>(props: CanvasProps<R, K> & { ref?: React.Ref<CanvasHandle> }) => JSX.Element;
+) as <R, K extends keyof R, SR>(props: CanvasProps<R, K, SR> & { ref?: React.Ref<CanvasHandle> }) => JSX.Element;