+++ src/Columns.tsx
@@ -4,10 +4,11 @@ import { Column } from './common/types';
 
 // TODO: fix type
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const SelectColumn: Column<any> = {
+export const SelectColumn: Column<any, any> = {
   key: 'select-row',
   name: '',
   width: 35,
+  maxWidth: 35,
   frozen: true,
   headerRenderer(props) {
     return (
@@ -18,7 +19,7 @@ export const SelectColumn: Column<any> = {
     );
   },
   formatter(props) {
-    return props.isSummaryRow ? null : (
+    return (
       <SelectCellFormatter
         value={props.isRowSelected}
         onChange={props.onRowSelectionChange}