+++ src/RowGroup.tsx
@@ -3,10 +3,10 @@ import React, { forwardRef } from 'react';
 import { RowExpandToggleEvent, Omit, CellRendererProps } from './common/types';
 import EventBus from './EventBus';
 
-interface Props<R> {
+interface Props<R, SR> {
   height: number;
   row: unknown;
-  cellRenderer?: React.ComponentType<CellRendererProps<R>>;
+  cellRenderer?: React.ComponentType<CellRendererProps<R, SR>>;
   isSelected?: boolean;
   rowIdx: number;
   forceUpdate?: boolean;
@@ -18,10 +18,10 @@ interface Props<R> {
   name: string;
   renderer?: React.ComponentType;
   eventBus: EventBus;
-  onRowExpandToggle?(event: RowExpandToggleEvent): void;
+  onRowExpandToggle?: (event: RowExpandToggleEvent) => void;
 }
 
-export default forwardRef<HTMLDivElement, Props<any>>(function RowGroup(props, ref) {
+export default forwardRef<HTMLDivElement, Props<any, any>>(function RowGroup(props, ref) {
   function onRowExpandToggle(expand?: boolean) {
     const { onRowExpandToggle } = props;
     if (onRowExpandToggle) {
@@ -41,15 +41,15 @@ export default forwardRef<HTMLDivElement, Props<any>>(function RowGroup(props, r
   const Renderer = props.renderer || DefaultBase;
 
   return (
-    <div onClick={onClick}>
+    <div className="rdg-row-group" onClick={onClick}>
       <Renderer {...props} ref={ref} onRowExpandClick={onRowExpandClick} onRowExpandToggle={onRowExpandToggle} />
     </div>
   );
-}) as <R>(props: Props<R> & { ref?: React.Ref<HTMLDivElement> }) => JSX.Element;
+}) as <R, SR>(props: Props<R, SR> & { ref?: React.Ref<HTMLDivElement> }) => JSX.Element;
 
-interface DefaultBaseProps extends Omit<Props<any>, 'onRowExpandToggle'> {
-  onRowExpandClick(): void;
-  onRowExpandToggle(expand?: boolean): void;
+interface DefaultBaseProps extends Omit<Props<any, any>, 'onRowExpandToggle'> {
+  onRowExpandClick: () => void;
+  onRowExpandToggle: (expand?: boolean) => void;
 }
 
 const DefaultBase = forwardRef<HTMLDivElement, DefaultBaseProps>(function DefaultBase(props, ref) {
@@ -71,7 +71,7 @@ const DefaultBase = forwardRef<HTMLDivElement, DefaultBaseProps>(function Defaul
 
   return (
     <div
-      className="rdg-row-group"
+      className="rdg-row-default-group"
       style={{ height }}
       onKeyDown={onKeyDown}
       tabIndex={0}