+++ src/SummaryCell.tsx
@@ -1,0 +1,45 @@
+import React, { memo } from 'react';
+import classNames from 'classnames';
+
+import { CellRendererProps } from './common/types';
+
+type SharedCellRendererProps<R, SR> = Pick<CellRendererProps<R, SR>,
+  | 'lastFrozenColumnIndex'
+  | 'scrollLeft'
+  | 'column'
+>;
+
+interface SummaryCellProps<R, SR> extends SharedCellRendererProps<R, SR> {
+  row: SR;
+}
+
+function SummaryCell<R, SR>({
+  column,
+  lastFrozenColumnIndex,
+  row,
+  scrollLeft
+}: SummaryCellProps<R, SR>) {
+  const { summaryFormatter: SummaryFormatter, width, left, summaryCellClass } = column;
+  const className = classNames(
+    'rdg-cell',
+    {
+      'rdg-cell-frozen': column.frozen,
+      'rdg-cell-frozen-last': column.idx === lastFrozenColumnIndex
+    },
+    typeof summaryCellClass === 'function' ? summaryCellClass(row) : summaryCellClass
+  );
+
+  const style: React.CSSProperties = { width, left };
+
+  if (scrollLeft !== undefined) {
+    style.transform = `translateX(${scrollLeft}px)`;
+  }
+
+  return (
+    <div className={className} style={style}>
+      {SummaryFormatter && <SummaryFormatter column={column} row={row} />}
+    </div>
+  );
+}
+
+export default memo(SummaryCell) as <R, SR>(props: SummaryCellProps<R, SR>) => JSX.Element;