+++ src/SummaryRowRenderer.tsx
@@ -1,6 +1,0 @@
-import { memo } from 'react';
-
-import { RowRendererProps } from './common/types';
-import Row from './Row';
-
-export default memo(Row) as <R>(props: RowRendererProps<R>) => JSX.Element;